requirements.txt
python-dotenv
langchain
langchain_openai
langchain-community
requests
langchain_google_genai